Refactor Component
